requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
bleach==6.1.0
Django==4.2.14
django-summernote==0.8.20.0
djangorestframework==3.15.2
six==1.16.0
sqlparse==0.5.1
typing-extensions==4.12.2
webencodings==0.5.1
